ecommerce_project/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-crispy-forms==2.3
django-widget-tweaks==1.5.0
idna==3.10
requests==2.32.3
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
